WIP: VDEV_BIAS_EXEMPT allocation class
The goal of this commit is to to add an allocation class that excludes
vdevs from being considered for metaslab allocation.

A new allocation class called "exempt" is added.
Devices with this allocation bias are never considered for metaslab allocation
and are consequently never written to, except for the vdev label range.
If a write to a location outside of the vdev label range is attempted,
we panic in the ZIO pipeline.

The idea behind VDEV_BIAS_EXEMPT is the following:
- Devices in that class are still regular leaf vdevs that support all the
  zpool management commands, show up in zpool status, prevent pool
  import if missing, etc.
- But all device space with the exception of the vdev label is available
  for other ZFS subsystems to use.

Note that this patch does not (yet) provide any mechanisms of coordination
between consumers of vdevs in the VDEV_BIAS_EXEMPT. We will need
mechanisms to:
- hold vdevs so that zpool remove does not work if some subsystem uses
  an exempt vdev
- persist ownership, e.g., have a mechanism so that a ZFS component
  can 'own' a vdev in the 'exempt' class and 'find' the devices
  it owns after reboot
 - ...?